Polished test runners
